+++ lib/CHANGELOG.md
@@ -33,6 +33,7 @@
    This is controlled with a new `posixly_correct` parser setting, disabled by default.
    Enforcement of “posixly correct” parsing behaviour means disallowing mixing of options and
    positionals, thus all arguments after a positional are to be interpreted as being positionals.
+ * Added methods to `OptionSetEx` for adding a pair of related long and short options together
  * Renamed the analysis `NonOption` variant to `Positional`, better distinguishing between
    *command non-options* and *positional non-options*.
  * Added an `add_shorts_from_str` method to `OptionSetEx`, which allows adding multiple short